Update 4/30/18 @ 11:12am
A huge fix in the Library & ViewLibrary classes. The program is now able to print a single book, all of the books, and I have started with getting the CheckOutBook class to work.

Next step:
Work on input verification for CheckOutBook class and create the methods/class for return book.